Create README.md
As a Web Developer, My objective is to make a
positive impact on clients, Co-workers and the
internet using my skills and experience to develop
compelling and attractive websites. Solving code
problems. Editing and developing with development
team in the company to build a perfect website.
Experienced web page designers with extensive
programming. Proficient in Larave Codeigniter
WordPress, HTML5, CSS, and responsive web
designs. Basic knowledge of Photoshop